added Message List, New Message, Pagination and Message Delete

+++ app/Controller/MessagesController.php
@@ -21,7 +21,16 @@ class MessagesController extends AppController {
  * @return void
  */
 	public function index() {
-		$this->Message->recursive = 0;
+		$this->Paginator->settings = array(
+			'conditions' => array(
+				'OR' => array(
+					array('Message.to_id' => $this->Auth->user('id')),
+					array('Message.from_id' => $this->Auth->user('id'))
+				)
+			),
+	        'order' => array('Message.created' => 'desc'),
+	        'limit' => 10
+	    );
 		$this->set('messages', $this->Paginator->paginate());
 	}
 
@@ -48,16 +57,25 @@ class MessagesController extends AppController {
 	public function add() {
 		if ($this->request->is('post')) {
 			$this->Message->create();
+			$this->request->data['Message']['from_id'] = $this->Auth->user('id');
 			if ($this->Message->save($this->request->data)) {
+				$this->request->data['MessageContent'] = $this->request->data['Message'];
+				$this->request->data['MessageContent']['message_id'] = $this->Message->id;
+				$this->loadModel('MessageContent');
+				$this->MessageContent->create();
+				$this->MessageContent->save($this->request->data);
 				$this->Flash->success(__('The message has been saved.'));
 				return $this->redirect(array('action' => 'index'));
-			} else {
-				$this->Flash->error(__('The message could not be saved. Please, try again.'));
 			}
 		}
-		$tos = $this->Message->To->find('list');
-		$froms = $this->Message->From->find('list');
-		$this->set(compact('tos', 'froms'));
+		$messageTos = $this->Message->MessageTo->find('list',
+			array(
+				'conditions' => array(
+					'MessageTo.id !=' => $this->Auth->user('id')
+				)
+			)
+		);
+		$this->set(compact('messageTos'));
 	}
 
 /**
@@ -67,7 +85,7 @@ class MessagesController extends AppController {
  * @param string $id
  * @return void
  */
-	public function edit($id = null) {
+	/*public function edit($id = null) {
 		if (!$this->Message->exists($id)) {
 			throw new NotFoundException(__('Invalid message'));
 		}
@@ -82,10 +100,10 @@ class MessagesController extends AppController {
 			$options = array('conditions' => array('Message.' . $this->Message->primaryKey => $id));
 			$this->request->data = $this->Message->find('first', $options);
 		}
-		$tos = $this->Message->To->find('list');
-		$froms = $this->Message->From->find('list');
-		$this->set(compact('tos', 'froms'));
-	}
+		$messageTos = $this->Message->MessageTo->find('list');
+		$messageFroms = $this->Message->MessageFrom->find('list');
+		$this->set(compact('messageTos', 'messageFroms'));
+	}*/
 
 /**
  * delete method
@@ -98,13 +116,20 @@ class MessagesController extends AppController {
 		$this->Message->id = $id;
 		if (!$this->Message->exists()) {
 			throw new NotFoundException(__('Invalid message'));
+			exit;
 		}
-		$this->request->allowMethod('post', 'delete');
+		$options = array('conditions' => array('Message.' . $this->Message->primaryKey => $id));
+		$message = $this->Message->find('first', $options);
+		if ($message['Message']['from_id'] != $this->Auth->user('id')) {
+			throw new NotFoundException(__('Invalid message'));
+			exit;
+		}
+
 		if ($this->Message->delete()) {
-			$this->Flash->success(__('The message has been deleted.'));
+			echo 'The message has been deleted.';
 		} else {
-			$this->Flash->error(__('The message could not be deleted. Please, try again.'));
+			echo 'The message could not be deleted. Please, try again.';
 		}
-		return $this->redirect(array('action' => 'index'));
+		exit;
 	}
 }
